requirements.txt
numpy==1.19.5
opencv-python==4.5.5.64
Pillow==8.4.0
